+++ src/main/java/domain/card/Number.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Number {
     A(1),
